Filterset tests.
