+++ src/tests/test_dynamic_api/test_filters.py
@@ -1,0 +1,73 @@
+from datetime import date
+
+import pytest
+from dso_api.dynamic_api.filterset import filterset_factory
+
+@pytest.mark.django_db
+class TestDynamicFilterSet:
+
+    @staticmethod
+    def test_serializer_has_nested_table(
+            parkeervakken_parkeervak_model, parkeervakken_regime_model
+    ):
+        """Prove that the serializer factory properly generates nested tables.
+        Serialiser should contain reverse relations.
+        """
+        parkeervak = parkeervakken_parkeervak_model.objects.create(
+            id="121138489047",
+            type="File",
+            soort="MULDER",
+            aantal=1.0,
+            e_type="E9",
+            buurtcode="A05d",
+            straatnaam="Zoutkeetsgracht",
+        )
+        parkeervakken_regime_model.objects.create(
+            id=1,
+            parent=parkeervak,
+            bord="",
+            dagen=["ma", "di", "wo", "do", "vr", "za", "zo"],
+            soort="MULDER",
+            aantal=None,
+            e_type="E6b",
+            kenteken="69-SF-NT",
+            opmerking="",
+            eind_tijd="23:59:00",
+            begin_tijd="00:00:00",
+            eind_datum=None,
+            begin_datum=None,
+        )
+        other_parkeervak = parkeervakken_parkeervak_model.objects.create(
+            id="121138489006",
+            type="File",
+            soort="MULDER",
+            aantal=1.0,
+            e_type="E6b",
+            buurtcode="A05d",
+            straatnaam="Zoutkeetsgracht",
+        )
+        parkeervakken_regime_model.objects.create(
+            id=2,
+            parent=other_parkeervak,
+            bord="",
+            dagen=["ma", "di", "wo", "do", "vr", "za", "zo"],
+            soort="MULDER",
+            aantal=None,
+            e_type="E6b",
+            kenteken="69-SF-NT",
+            opmerking="",
+            eind_tijd="23:59:00",
+            begin_tijd="00:00:00",
+            eind_datum=None,
+            begin_datum=None,
+        )
+
+        filterset_class = filterset_factory(parkeervakken_parkeervak_model)
+        filterset = filterset_class({"regimes.eType": "E6b"})
+        breakpoint()
+
+        assert filterset.is_valid(), filterset.errors
+
+        result = filterset.filter_queryset(parkeervakken_parkeervak_model.objects.all())
+        assert result.count() == 1, result
+        breakpoint()
